+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 * middleman-blog に立脚スタイルに戻す
   * 車輪の再発明すぎたし
+* category manager と主要 helper や template, partials をいれて、とりあえずエラーなく動くように  
 
 ### 0.0.7